Rework handoff to eliminate double-counting and clean up dead code
Sender-side visual delay: workers approaching an edge are marked as
"departing" — sprite vanishes and movement freezes for 2s while the
worker remains owned by the sending module. After the delay, the
actual ESP-NOW transfer fires. Receiver places workers immediately
(no tunnel buffer delay). This keeps each worker in exactly one place
at all times, fixing population count fluctuation during handoffs.

Population display throttled to 5-second updates to smooth any
remaining transit jitter.

Code review cleanup:
- Remove TunnelPending buffer (dead code since receiver is instant)
- Simplify _receive_handoffs to place directly via _place_arrival
- Widen handoff seq uint8 -> uint16 (65K before wrap vs 256)
- Widen HandoffAck seq and dedup table to match
- Increase ACK buffer 16 -> 32 (handles burst handoffs)
- Clean up ESP-NOW peer on face disconnect (prevents peer table leak)
- Satellite pop broadcast interval 1s -> 5s (less redundant traffic)
- Skip departing workers in chamber tick and renderer

+++ firmware/include/topology.h
@@ -117,7 +117,7 @@ struct BoundaryPheroData {
 struct __attribute__((packed)) HandoffAck {
     uint8_t  msg_type;    // TOPO_HANDOFF_ACK
     uint16_t acker_id;    // module that received the ant
-    uint8_t  seq;         // matches seq in LilGuyTransfer
+    uint16_t seq;         // matches seq in LilGuyTransfer
 };
 
 // Send raw payload to a face's neighbour (for handoff transfers)
